test: separate getContract tests from SafeProvider tests
